+++ core/common/src/main/kotlin/com/kelsos/mbrc/core/common/state/AppState.kt
@@ -8,6 +8,7 @@ interface AppStateFlow {
   val playerStatus: StateFlow<PlayerStatusModel>
   val playingTrack: StateFlow<TrackInfo>
   val playingTrackRating: StateFlow<TrackRating>
+  val playingTrackDetails: StateFlow<TrackDetails>
   val playingPosition: StateFlow<PlayingPosition>
   val lyrics: StateFlow<List<String>>
 }
@@ -19,6 +20,8 @@ interface AppStatePublisher : AppStateFlow {
 
   fun updateTrackRating(rating: TrackRating)
 
+  fun updateTrackDetails(details: TrackDetails)
+
   fun updatePlayingPosition(position: PlayingPosition)
 
   fun updateLyrics(lyrics: List<String>)
@@ -28,12 +31,14 @@ class AppState : AppStatePublisher {
   private val _playerStatus = MutableStateFlow(PlayerStatusModel())
   private val _playingTrack = MutableStateFlow<TrackInfo>(BasicTrackInfo())
   private val _playingTrackRating = MutableStateFlow(TrackRating())
+  private val _playingTrackDetails = MutableStateFlow(TrackDetails.EMPTY)
   private val _playingPosition = MutableStateFlow(PlayingPosition())
   private val _lyrics = MutableStateFlow(emptyList<String>())
 
   override val playerStatus: StateFlow<PlayerStatusModel> = _playerStatus.asStateFlow()
   override val playingTrack: StateFlow<TrackInfo> = _playingTrack.asStateFlow()
   override val playingTrackRating: StateFlow<TrackRating> = _playingTrackRating.asStateFlow()
+  override val playingTrackDetails: StateFlow<TrackDetails> = _playingTrackDetails.asStateFlow()
   override val playingPosition: StateFlow<PlayingPosition> = _playingPosition.asStateFlow()
   override val lyrics: StateFlow<List<String>> = _lyrics.asStateFlow()
 
@@ -47,6 +52,10 @@ class AppState : AppStatePublisher {
 
   override fun updateTrackRating(rating: TrackRating) {
     _playingTrackRating.value = rating
+  }
+
+  override fun updateTrackDetails(details: TrackDetails) {
+    _playingTrackDetails.value = details
   }
 
   override fun updatePlayingPosition(position: PlayingPosition) {