+++ core/networking/src/main/kotlin/com/kelsos/mbrc/core/networking/api/PlaybackApiImpl.kt
@@ -5,6 +5,7 @@ import com.kelsos.mbrc.core.networking.ApiBase
 import com.kelsos.mbrc.core.networking.dto.NowPlayingDto
 import com.kelsos.mbrc.core.networking.protocol.base.Protocol
 import com.kelsos.mbrc.core.networking.protocol.payloads.CoverPayload
+import com.kelsos.mbrc.core.networking.protocol.payloads.NowPlayingDetailsPayload
 import kotlinx.coroutines.flow.Flow
 
 class PlaybackApiImpl(private val apiBase: ApiBase) : PlaybackApi {
@@ -13,4 +14,7 @@ class PlaybackApiImpl(private val apiBase: ApiBase) : PlaybackApi {
 
   override suspend fun getCover(): CoverPayload =
     apiBase.getItem(Protocol.NowPlayingCover, CoverPayload::class)
+
+  override suspend fun getTrackDetails(): NowPlayingDetailsPayload =
+    apiBase.getItem(Protocol.NowPlayingDetails, NowPlayingDetailsPayload::class)
 }